requirements.txt
asgiref==3.7.2
Django==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
djangorestframework==3.12.2
ipdb==0.13.13
# psycopg2-binary==2.9.3